+ creating tasks in iteration page + changed assignment availability data type from short to integer
git-svn-id: http://agilefant.googlecode.com/svn/trunk@2810 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/business/AssignmentBusiness.java
@@ -9,8 +9,8 @@ import fi.hut.soberit.agilefant.model.ExactEstimate;
 
 public interface AssignmentBusiness extends GenericBusiness<Assignment> {
     public Assignment store(int assignmentId, ExactEstimate personalLoad,
-            short availability);
+            int availability);
 
     public Collection<Assignment> addMultiple(Backlog backlog,
-            Set<Integer> userIds, ExactEstimate personalLoad, short availability);
+            Set<Integer> userIds, ExactEstimate personalLoad, int availability);
 }
